Fix generator for recursive structures.
As we are processing a struct, it might be that we hit that same struct again as a child. We can't process twice the same struct without falling into an infinite loop.

This change detects such case. It defers the point at which structs get their subfields until the end, because we don't know all of them beforehand, and for that we need an extra layer of indirection for the fields slice, because slice are immutable.